API/requirements.txt
fastapi
uvicorn
sqlalchemy
PyJWT
passlib
python-multipart